+++ package.json
@@ -9,39 +9,23 @@
     "@fontsource/roboto": "^5.0.8",
     "@mui/icons-material": "^5.15.9",
     "@mui/material": "^5.15.9",
-    "@testing-library/jest-dom": "^6.4.2",
-    "@testing-library/react": "^14.2.1",
-    "@testing-library/user-event": "^14.5.2",
-    "@types/jest": "^29.5.12",
     "@types/node": "^20.11.17",
     "@types/react": "^18.2.55",
     "@types/react-dom": "^18.2.19",
     "axios": "^1.6.7",
-    "css-select": "5.1.0",
-    "css-selector": "^0.1.0",
     "gh-pages": "^6.1.1",
-    "postcss": "^8.4.31",
+    "next": "^14.2.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.22.0",
-    "react-scripts": "5.0.1",
-    "typescript": "^5.1.0",
-    "web-vitals": "^3.5.2",
-    "webpack-dev-middleware": "5.3.4"
+    "typescript": "^5.1.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "predeploy": "yarn run build",
-    "deploy": "gh-pages -d build",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint",
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d out"
   },
   "browserslist": {
     "production": [
