+++ pages/podcasts.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { Typography } from "@mui/material";
+
+export default function Podcasts() {
+  return (
+    <div>
+      <Typography variant="h3" gutterBottom sx={{ fontWeight: "bold" }}>
+        Podcasts
+      </Typography>
+      <main>
+        <Typography variant="body1" gutterBottom>
+          Coming soon!
+        </Typography>
+      </main>
+    </div>
+  );
+}
