+++ src/types.d.ts
@@ -1,0 +1,14 @@
+declare module "*.jpeg" {
+  const value: any;
+  export default value;
+}
+
+declare module "*.jpg" {
+  const value: any;
+  export default value;
+}
+
+declare module "*.png" {
+  const value: any;
+  export default value;
+}
